+++ source/extensions/lightspeed.trex.stage_manager.plugin.tree.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.1]
+### Changed
+- Made mesh grouping always use parent/child hierarchy labeling for children
+
 ## [1.2.0]
 ### Fixed
 - Fixed issue where a prim could not be added to more than 1 category